+++ tower-ready-cache/Cargo.toml
@@ -29,3 +29,7 @@ log = "0.4.1"
 tokio-sync = "0.1.3"
 tower-service = "0.2.0"
 tower-util = "0.1.0"
+
+[dev-dependencies]
+tokio-executor = "0.1.7"
+tower-test = { version = "0.1.0", path = "../tower-test" }
